refact: re-organize versions endpoint
For the API, move the `versions.rs` into `versions/mod.rs`. The two
modules for request and response are now in `request/mod.rs` and
`response/mod.rs`. Historical messages are in the `response/messages.rs`
module
